+++ lib/world/events.ts
@@ -18,7 +18,7 @@ export class WorldAfterEvents extends Wrapper<MC.WorldAfterEvents> {
     public readonly entityDie:        IEventSignal<EntityDieAfterEvent, EntityEventOptions>;
     public readonly itemUse:          IEventSignal<ItemUseAfterEvent>;
     public readonly pistonActivate:   IEventSignal<PistonActivateAfterEvent>;
-    public readonly worldInitialize:  IEventSignal<MC.WorldInitializeAfterEvent>;
+    public readonly worldLoad:        IEventSignal<MC.WorldLoadAfterEvent>;
     /** An event that is fired when the world is fully loaded. */
     public readonly ready:            CustomEventSignal<ReadyAfterEvent>;
     public readonly playerSpawn:      CustomEventSignal<PlayerSpawnAfterEvent>;
@@ -72,7 +72,7 @@ export class WorldAfterEvents extends Wrapper<MC.WorldAfterEvents> {
         this.pistonActivate = new GluedEventSignalWithoutOptions(
             this.raw.pistonActivate,
             (rawEv: MC.PistonActivateAfterEvent) => new PistonActivateAfterEvent(rawEv));
-        this.worldInitialize = this.raw.worldInitialize;
+        this.worldLoad   = this.raw.worldLoad;
         this.ready       = new CustomEventSignal();
         this.playerSpawn = new CustomEventSignal();
         this.playerLeave = new CustomEventSignal();